Lazada Review Data Analysis
Lazada is an international e-commerce company and one of the largest e-commerce operators in Southeast Asia. It was founded in 2012 by Maximilian Bittner with the intention of establishing an Amazon-like business model in the region. Lazada operates in six countries: Indonesia, Malaysia, the Philippines, Singapore, Thailand, and Vietnam. It offers a wide range of products across various categories, such as fashion, electronics, home and living, mother and baby, and health and beauty. It also has a virtual mall called LazMall, where consumers can shop from leading international and local brands.